contrib: add a pair of credential helpers for Mac OS X's keychain
This credential helper adds, searches, and removes entries from
the Mac OS X keychain. The C version links against the Security
framework and is probably the best choice for daily use.

A python version is also included primarily as a more readable
example and uses the /usr/bin/security CLI to access the keychain.

Tested with 10.6.8.

Signed-off-by: Jay Soffian <jaysoffian@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>