check_expirations: don't copy over same element
The credentials expiry loop looks at a credential, and if it's
expired, frees it and memcpy()s the last credential in the now free
place.

This results in a memcpy() with source=destination, which technically
yields undefined behaviour.  Instead of turning it into a memmove,
don't copy anything if we deleted the last entry.

Signed-off-by: Thomas Rast <trast@student.ethz.ch>
Acked-by: Jeff King <peff@peff.net>
Signed-off-by: Junio C Hamano <gitster@pobox.com>